requirements.txt
langchain>=0.2.0
openai>=1.0.0
faiss-cpu>=1.8.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
